requirements.txt
fastapi~=0.110.1
uvicorn[standard]~=0.29.0
psycopg2-binary~=2.9.9